レビュー機能を修正したり汎用性高く書き直しました
