+++ app/controllers/restaurants/reviews_controller.rb
@@ -1,0 +1,16 @@
+module Restaurants
+  class ReviewsController < ApplicationController
+    def create
+      restaurant = Restaurant.find(params[:restaurant_id])
+      review = restaurant.reviews.new(review_params)
+      review.user = current_user
+      review.save
+      redirect_to restaurant_url(restaurant)
+    end
+
+    private
+      def review_params
+        params.require(:review).permit(:content, :score)
+      end
+  end
+end
