+++ app/models/restaurant.rb
@@ -1,8 +1,4 @@
 class Restaurant < ApplicationRecord
   belongs_to :category
   has_many :reviews
-
-  def reviews_with_id
-    reviews.reviews_with_id
-  end
 end
