+++ app/models/review.rb
@@ -2,14 +2,6 @@ class Review < ApplicationRecord
   belongs_to :restaurant
   belongs_to :user
 
-  scope :reviews_with_id, -> { where.not(restaurant_id: nil) }
-  scope :star_repeat_select, -> {
-    {
-      "★★★★★" => 5,
-      "★★★★"  => 4,
-      "★★★"   => 3,
-      "★★"    => 2,
-      "★"     => 1
-    }
-  }
+  enum score: { five: 5, four: 4, three: 3, two: 2, one: 1 }
+
 end
