+++ app/views/restaurants/show.html.erb
@@ -8,7 +8,7 @@
     <th>住所</th>
     <th>電話番号</th>
     <th>定休日</th>
-    <th>カテゴリID</th>
+    <th>カテゴリ</th>
   </tr>
   <tr>
     <td><%= @restaurant.name %></td>
@@ -24,23 +24,27 @@
 </table>
 
 <div class="w-50 mt-5">
-<p class="fs-6 fw-bold">レビュー</p>
+  <p class="fs-6 fw-bold">レビュー</p>
   <% @reviews.each do |review| %>
-  <div class="">
-    <p class="review-score-color"><%= score_on_star(review.score) %></p>
-    <p class="fs-7"><%= review.content %></p>
-    <p class="fs-7"><%= review.user.name %></p>
-    <label><%= review.created_at.strftime('%Y/%m/%d') %></label>
-  </div>
+    <div>
+      <p class="review-score-color">
+        <% Review.scores[review.score].to_i.times.each do |i| %>
+          ★
+        <% end %>
+      </p>
+      <p class="fs-7"><%= review.content %></p>
+      <p class="fs-7"><%= review.user.name %></p>
+      <p><%= l review.created_at, format: :normal %></p>
+    </div>
   <% end %>
 </div>
 
 <% if user_signed_in? %>
   <div class="w-50">
-    <div class="">
-      <%= form_with model: @review, url: restaurant_reviews_path(@restaurant) do |f| %>
-      <p class="fs-6 fw-bold">評価</p>
-        <%= f.select :score, @star_repeat_select, {}, class: "form-control m-2 review-score-color" %>
+    <div>
+      <%= form_with model: Review.new, url: restaurant_reviews_path(@restaurant) do |f| %>
+        <p class="fs-6 fw-bold">評価</p>
+        <%= f.select :score, Review.scores_i18n.keys.map {|key| [I18n.t("activerecord.enums.review.score.#{key}"), key]}, {}, class: "form-control m-2 review-score-color" %>
         <p class="fs-6 fw-bold">レビュー内容</p>
         <%= f.text_area :content, class: "form-control m-2" %>
         <%= f.submit "レビューを追加" , class:"btn nagoyameshi-submit-button ml-2"%>
