+++ db/migrate/20240403060939_create_reviews.rb
@@ -1,6 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[7.1]
   def change
     create_table :reviews do |t|
+      # レビューの内容
       t.text :content
       t.references :restaurant, null: false, foreign_key: true
       t.references :user, null: false, foreign_key: true
